onClick BetPlace -> count #bet, total amount, total win

+++ database/migrations/2020_08_17_034945_create_answers_table.php
@@ -21,8 +21,8 @@ class CreateAnswersTable extends Migration
             $table->float('bet_rate');
 
             $table->unsignedBigInteger('place')->default(0);
-            $table->unsignedBigInteger('bet_amount')->default(0);
-            $table->unsignedBigInteger('rtrn_amount')->default(0);
+            $table->double('bet_amount')->default(0);
+            $table->double('rtrn_amount')->default(0);
             $table->enum('status', ['active', 'inactive', 'win', 'loss'])->default('active');
             $table->enum('result', ['winned', 'lossed', ''])->default('');
             $table->timestamps();
